on-handoff-orchestrator-agent/pyproject.toml
[project]
name = "handoff-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
]
